python_advanced_diploma/pyproject.toml
[tool.poetry]
name = "python-advanced-diploma"
version = "0.1.0"
description = "This is a Python advanced diploma project"
authors = ["nikolasy <bogachev.pro@gmail.com>"]
license = "Mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
databases = "^0.9.0"
yadisk = "^3.1.0"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"



[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
